refactor: rename cookie popup to cookie box
