+++ packages/ui/public/index.html
@@ -35,7 +35,6 @@
       You need to enable JavaScript to run this app.
     </noscript>
     <div id="root"></div>
-    <div id="cookie-popup-root"></div>
     <!--
       This HTML file is a template.
       If you open it directly in the browser, you will see an empty page.
